errors fixed

+++ pages/settings_general/update.php
@@ -6,7 +6,7 @@ $settings = new Settings();
 // check for method
 if (Helpers::getRequestMethod() == "GET") {
     // get values from database to show them in inputs fields
-    $stmt = $pdo->prepare("SELECT company, slogan, keywords, site_title, site_url, smtp_url, smtp_port, 
+    $stmt = $pdo->prepare("SELECT company, slogan, description, keywords, site_title, site_url, smtp_url, smtp_port, 
     normal_photo, normal_photo_width, normal_photo_height, top_photo, top_photo_width, top_photo_height, 
     small_photo, small_photo_width, small_photo_height, debug_mode, maintenance_mode, maintenance_mode_content
     FROM settings
@@ -27,6 +27,7 @@ if (Helpers::getRequestMethod() == "GET") {
     // show values
     $settings->company->value = $result['company'];
     $settings->slogan->value = $result['slogan'];
+    $settings->description->value = $result['description'];
     $settings->keywords->value = $result['keywords'];
     $settings->site_title->value = $result['site_title'];
     $settings->site_url->value = $result['site_url'];
@@ -58,14 +59,15 @@ if (Helpers::getRequestMethod() == 'POST') {
     $result = $stmt->fetch(PDO::FETCH_ASSOC);
     $stmt->closeCursor();
 
-    $normal_photo_exists = $result['normal_photo'];
-    $top_photo_exists = $result['top_photo'];
-    $small_photo_exists = $result['small_photo'];
+    $normal_photo_exists = $result['normal_photo'] ?? '';
+    $top_photo_exists = $result['top_photo'] ?? '';
+    $small_photo_exists = $result['small_photo'] ?? '';
 
     // grab data from form inputs
 
     $settings->company->value = htmlspecialchars($_POST[$settings->company->name] ?? '');
     $settings->slogan->value = htmlspecialchars($_POST[$settings->slogan->name] ?? '');
+    $settings->description->value = htmlspecialchars($_POST[$settings->description->name] ?? '');
     $settings->keywords->value = htmlspecialchars($_POST[$settings->keywords->name] ?? '');
     $settings->site_title->value = htmlspecialchars($_POST[$settings->site_title->name] ?? '');
     $settings->site_url->value = htmlspecialchars($_POST[$settings->site_url->name] ?? '');
@@ -96,13 +98,15 @@ if (Helpers::getRequestMethod() == 'POST') {
 
             // delete old file
             if ($normal_photo_status['code']) {
-                unlink(ltrim($normal_photo_exists, "/"));
+                if (file_exists(ltrim($normal_photo_exists, "/"))) {
+                    unlink(ltrim($normal_photo_exists, "/"));
+                }
             }
 
             // assign values from upload function
             $normal_photo_check = $normal_photo_status['code'];
-            $users->normal_photo->value = $normal_photo_status['file'];
-            $users->normal_photo->error_msg = $normal_photo_status['msg'];
+            $settings->normal_photo->value = $normal_photo_status['file'];
+            $settings->normal_photo->error_msg = $normal_photo_status['msg'];
         }
     } else {
         // if no photo was attempted to be uploaded then assign its value
@@ -124,13 +128,15 @@ if (Helpers::getRequestMethod() == 'POST') {
 
             // delete old file
             if ($top_photo_status['code']) {
-                unlink(ltrim($top_photo_exists, "/"));
+                if (file_exists(ltrim($top_photo_exists, "/"))) {
+                    unlink(ltrim($top_photo_exists, "/"));
+                }
             }
 
             // assign values from upload function
             $top_photo_check = $top_photo_status['code'];
-            $users->top_photo->value = $top_photo_status['file'];
-            $users->top_photo->error_msg = $top_photo_status['msg'];
+            $settings->top_photo->value = $top_photo_status['file'];
+            $settings->top_photo->error_msg = $top_photo_status['msg'];
         }
     } else {
         // if no photo was attempted to be uploaded then assign its value
@@ -152,13 +158,15 @@ if (Helpers::getRequestMethod() == 'POST') {
 
             // delete old file
             if ($small_photo_status['code']) {
-                unlink(ltrim($small_photo_exists, "/"));
+                if (file_exists(ltrim($small_photo_exists, "/"))) {
+                    unlink(ltrim($small_photo_exists, "/"));
+                }
             }
 
             // assign values from upload function
             $small_photo_check = $small_photo_status['code'];
-            $users->small_photo->value = $small_photo_status['file'];
-            $users->small_photo->error_msg = $small_photo_status['msg'];
+            $settings->small_photo->value = $small_photo_status['file'];
+            $settings->small_photo->error_msg = $small_photo_status['msg'];
         }
     } else {
         // if no photo was attempted to be uploaded then assign its value
@@ -174,8 +182,8 @@ if (Helpers::getRequestMethod() == 'POST') {
 
     // check if given data is ok
     $checks = $settings->company->check() && $settings->slogan->check() && $settings->keywords->check() && 
-                $settings->site_title->check() && $settings->site_url->check() && $settings->smtp_url->check() && 
-                $settings->smtp_password->check() && $settings->smtp_port->check() && 
+                $settings->description->check() && $settings->site_title->check() && $settings->site_url->check() && 
+                $settings->smtp_url->check() && $settings->smtp_password->check() && $settings->smtp_port->check() && 
                 $settings->normal_photo_width->check() && $settings->normal_photo_height->check() && 
                 $settings->top_photo_width->check() && $settings->top_photo_height->check() && 
                 $settings->small_photo_width->check() && $settings->small_photo_height->check() && $settings->debug_mode->check() && 
@@ -186,7 +194,7 @@ if (Helpers::getRequestMethod() == 'POST') {
         $updated_at = date($datetime_format, $settings->updated_at->value->getTimestamp());
 
         // sql statement
-        $stmt = $pdo->prepare("UPDATE settings SET company = :company, slogan = :slogan, keywords = :keywords,
+        $stmt = $pdo->prepare("UPDATE settings SET company = :company, slogan = :slogan, description = :description, keywords = :keywords,
                                 site_title = :site_title, site_url = :site_url, smtp_url = :smtp_url,
                                 smtp_password = :smtp_password, smtp_port = :smtp_port, normal_photo = :normal_photo,
                                 normal_photo_width = :normal_photo_width, normal_photo_height = :normal_photo_height,
@@ -199,6 +207,7 @@ if (Helpers::getRequestMethod() == 'POST') {
         //  bind values and parameters
         $stmt->bindParam(':company', $settings->company->value, PDO::PARAM_STR);
         $stmt->bindParam(':slogan', $settings->slogan->value, PDO::PARAM_STR);
+        $stmt->bindParam(':description', $settings->description->value, PDO::PARAM_STR);
         $stmt->bindParam(':keywords', $settings->keywords->value, PDO::PARAM_STR);
         $stmt->bindParam(':site_title', $settings->site_title->value, PDO::PARAM_STR);
         $stmt->bindParam(':site_url', $settings->site_url->value, PDO::PARAM_STR);
